Module minwin::named

Structs

NamedOpenOptions
Options for opening a named object.

Enums

CreateNamedError

Traits

NamedBuilder
A trait containing common methods for creating named objects.

NamedObject
A type that can be constructed given an existing object name.

Type Definitions

CreateNamedResult

NamedOpenFunction
Windows API function type used for opening named objects.