Trait minwin::named::NamedObject

pub trait NamedObject: ObjectExt {
    fn open_function() -> NamedOpenFunction;
    fn default_open_options() -> NamedOpenOptions;

    fn open<N: AsRef<OsStr>>(name: N) -> Result<Self> { ... }
    fn open_with_options<N, O>(name: N, options: O) -> Result<Self> where N: AsRef<OsStr>, O: Into<NamedOpenOptions> { ... }
}

A type that can be constructed given an existing object name.

Required Methods

fn open_function() -> NamedOpenFunction

The extern function used for opening an object of this type.

fn default_open_options() -> NamedOpenOptions

The default opening options for this type.

Provided Methods

fn open<N: AsRef<OsStr>>(name: N) -> Result<Self>

Opens a named object using default options.

fn open_with_options<N, O>(name: N, options: O) -> Result<Self> where N: AsRef<OsStr>, O: Into<NamedOpenOptions>

Opens a named object using the specified options.