Struct minwin::sync::EventBuilder

pub struct EventBuilder {
    // some fields omitted
}

A builder for creating a new Event.

See CreateEvent and CreateEventEx for more information.

Methods

impl EventBuilder

fn new() -> EventBuilder

Creates a new event builder with the default settings.

fn manual_reset(&mut self, manual_reset: bool) -> &mut EventBuilder

The reset mode of the event.

If the parameter is true, the builder creates a manual-reset event object, which requires the use of the Event::reset() function to set the event state to nonsignaled. If this parameter is false, the function creates an auto-reset event object, and system automatically resets the event state to nonsignaled after a single waiting thread has been released.

fn initial_state(&mut self, initial_state: bool) -> &mut EventBuilder

The initial state of the event.

If the parameter is true, the initial state of the event object is signaled; otherwise, it is nonsignaled.

fn desired_access<A: Access>(&mut self, desired_access: A) -> &mut EventBuilder

The desired access for the event object.